fix: resolve lint unparam warnings and Windows build failure
Lint (unparam):
- git.go: suppress runWithEnv string return warning (kept for API consistency)
- witness/handlers.go: blank unused workDir and rigName params in
  handlePolecatDoneNoMR and handleMergedCleanupStatus (persistent polecat
  model removed their usage)

Windows build:
- Split doltserver.go into platform-specific files using build tags
- doltserver.go: platform-independent types, vars, and helpers
- doltserver_unix.go: syscall.Flock-based locking (Linux/macOS)
- doltserver_windows.go: lock-free fallback for Windows CI
